feat: Update the invocation_model to record invocation_status
Each record in the _executions tables have a status on whether they were a success or failure. The individual records were aggregated to determine whether the invocation was a success or a failure.

Additionally, the selection criteria is extracted out of the invocation_args into its own column. This combined with the invocation_status makes each record in the invocation table a complete picture of the dbt exectuion.